requirements.minimal.txt
discord.py
python-dotenv
aiohttp
psycopg2-binary
beautifulsoup4
Brotli
